services/core/pyproject.toml
[project]
name = "bush-core"
version = "0.1.0"
description = "Lightweight MQTT services (audio-agent, t2v-bridge, tts)"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0",
    "bushutil",
]

[project.scripts]
bush-audio-agent = "bush_audio_agent:main"
bush-t2v = "bush_t2v:main"
bush-tts = "bush_tts:main"
bush-flame-expression = "bush_flame_expression:main"

[tool.uv.sources]
bushutil = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/bush_audio_agent", "src/bush_t2v", "src/bush_tts", "src/bush_flame_expression"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
